chore: code clean up and files relocation
