+++ src/components/dashboard/pages/AddStakePage.tsx
